+++ README.md
@@ -1,1 +1,1 @@
-# LightningBringer
+# Lighter
