+++ docs/index.md
@@ -1,1 +1,1 @@
-# lightningbringer
+# lighter
