+++ lighter/__init__.py
@@ -1,3 +1,3 @@
-from .system import System
+from .system import LighterSystem
 from .config import init_config
 from .runner import init_system, init_trainer
