+++ lighter/utils.py
