+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: lightningbringer
+site_name: lighter
   
 nav:
     - Home: index.md
