+++ projects/nlst/experiments/gender.yaml
@@ -1,10 +1,10 @@
-project: /home/ibro/Projects/lightningbringer/projects/nlst
+project: /home/ibro/Projects/lighter/projects/nlst
 
 trainer:
     _target_: pytorch_lightning.Trainer
     benchmark: True
     max_epochs: 100
-    check_val_every_n_epoch: 5
+    check_val_every_n_epoch: 1
     accelerator: gpu
     devices: 1
     # strategy:
@@ -13,7 +13,7 @@ trainer:
     precision: 16
     log_every_n_steps: 50
     # logger:
-    #     _target_: lightningbringer.logger.LightningBringerLogger
+    #     _target_: lighter.logger.LighterLogger
     #     save_dir: ${project}/logs/${now:}
     #     tensorboard: True
     #     wandb: True
@@ -29,7 +29,7 @@ trainer:
     #       save_on_train_epoch_end: True
 
 system:
-    _target_: lightningbringer.System
+    _target_: lighter.LighterSystem
     batch_size: 8
     pin_memory: True
     num_workers: 3
@@ -66,8 +66,14 @@ system:
         weight_decay: 1e-5
 
     schedulers:
-        _target_: torch.optim.lr_scheduler.ExponentialLR
-        gamma: 0.95
+        # _target_: torch.optim.lr_scheduler.ExponentialLR
+        # gamma: 0.95
+    
+        scheduler:
+            _target_: torch.optim.lr_scheduler.ReduceLROnPlateau
+        interval: epoch
+        monitor: val/loss
+        strict: True
 
     train_metrics:
         - _target_: torchmetrics.Precision
