+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
-name = lightningbringer
+name = lighter
 version = 0.0.1
 author = "ibro45"
 # author-email = 
-url = https://github.com/ibro45/lightningbringer
-description = lightningbringer description
+url = https://github.com/ibro45/lighter
+description = lighter description
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = mit
@@ -15,7 +15,7 @@ classifiers =
     Programming Language :: Python :: 3 :: Only
     Intended Audience :: Science/Research
 project_urls =
-    Documentation = https://lightningbringer.readthedocs.io/en/latest/
+    Documentation = https://lighter.readthedocs.io/en/latest/
 
 [options]
 zip_safe = False
@@ -57,7 +57,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    lightningbringer = lightningbringer.cli:interface
+    lighter = lighter.cli:interface
 
 [yapf]
 based_on_style = google
